move files into original folders
