+++ src/_common/popper/popper.ts
@@ -1,196 +1,392 @@
-// import Vue from 'vue';
-// import { Component, Emit, Prop } from 'vue-property-decorator';
-// import { Backdrop } from '../backdrop/backdrop.service';
-// import AppBackdrop from '../backdrop/backdrop.vue';
-// import { Screen } from '../screen/screen-service';
-// import AppScrollScrollerTS from '../scroll/scroller/scroller';
-// import AppScrollScroller from '../scroll/scroller/scroller.vue';
-// import { Popper } from './popper.service';
-// import './popper.styl';
+import arrow, { ArrowModifier } from '@popperjs/core/lib/modifiers/arrow';
+import flip from '@popperjs/core/lib/modifiers/flip';
+import hide from '@popperjs/core/lib/modifiers/hide';
+import preventOverflow, {
+	PreventOverflowModifier,
+} from '@popperjs/core/lib/modifiers/preventOverflow';
+import { createPopper, Instance, Options } from '@popperjs/core/lib/popper-lite';
+import ResizeObserver from 'resize-observer-polyfill';
+import Vue from 'vue';
+import { Component, Emit, Prop, Watch } from 'vue-property-decorator';
+import { propOptional } from '../../utils/vue';
+import { Backdrop } from '../backdrop/backdrop.service';
+import AppBackdrop from '../backdrop/backdrop.vue';
+import { Screen } from '../screen/screen-service';
+import AppScrollScroller from '../scroll/scroller/scroller.vue';
+import { Popper } from './popper.service';
+import './popper.styl';
 
-// const mod: any = require('v-tooltip');
+// Sync with the styles files.
+const TransitionTime = 200;
 
-// // Sync with the styles files.
-// const TransitionTime = 200;
+let PopperIndex = 0;
 
-// let PopperIndex = 0;
+type ActualTrigger = 'click' | 'hover' | 'manual';
 
-// type ActualTrigger = 'click' | 'hover' | 'manual';
+type DelayFormat = {
+	show: number;
+	hide: number;
+};
 
-// @Component({
-// 	components: {
-// 		VPopover: mod.VPopover,
-// 		AppScrollScroller,
-// 	},
-// })
-// export default class AppPopper extends Vue {
-// 	@Prop({ type: String, default: 'bottom' })
-// 	placement!: 'top' | 'right' | 'bottom' | 'left';
+const modifiers = [
+	flip,
+	preventOverflow,
+	arrow,
+	hide,
+	{
+		// padding between popper and viewport
+		name: 'preventOverflow',
+		options: {
+			padding: 5,
+		},
+	} as PreventOverflowModifier,
+	{
+		// padding between popper-arrow and corner, to prevent overflow with border-radius
+		name: 'arrow',
+		options: {
+			padding: 8,
+		},
+	} as ArrowModifier,
+];
 
-// 	@Prop({ type: String, default: 'click' })
-// 	trigger!: ActualTrigger | 'right-click';
+@Component({
+	components: {
+		AppScrollScroller,
+	},
+})
+export default class AppPopper extends Vue {
+	@Prop(propOptional(String, 'bottom'))
+	placement!: 'top' | 'right' | 'bottom' | 'left';
 
-// 	/**
-// 	 * By default the popper will stay on the page until the user clicks outside
-// 	 * of the popper. This tells the popper to close anytime the state changes.
-// 	 * Useful for poppers in the shell that link to other pages on the site.
-// 	 */
-// 	@Prop(Boolean)
-// 	hideOnStateChange?: boolean;
+	@Prop(propOptional(String, 'click'))
+	trigger!: ActualTrigger | 'right-click';
 
-// 	/**
-// 	 * Whether or not the popper should size itself to the same width as the
-// 	 * trigger. Useful for poppers that work like "select" type controls.
-// 	 */
-// 	@Prop(Boolean)
-// 	trackTriggerWidth?: boolean;
+	/**
+	 * We want the popper to be 'display: fixed' if we use it on a fixed parent.
+	 * This should prevent stuttering on scroll if the popper is attached to the nav.
+	 */
+	@Prop(propOptional(Boolean))
+	fixed?: boolean;
 
-// 	/**
-// 	 * Whether or not the popper should take up the full max width instead of
-// 	 * relying on its content to size itself. Useful for poppers that change the
-// 	 * content dynamically and you want it to stay one consistent size.
-// 	 */
-// 	@Prop(Boolean)
-// 	forceMaxWidth?: boolean;
+	/**
+	 * By default the popper will stay on the page until the user clicks outside
+	 * of the popper. This tells the popper to close anytime the state changes.
+	 * Useful for poppers in the shell that link to other pages on the site.
+	 *
+	 * Does not work on 'hover' poppers, so instead we run
+	 * 'popperInstance.destroy()' in the 'destroyed()' lifecycle.
+	 */
+	@Prop(propOptional(Boolean))
+	hideOnStateChange?: boolean;
 
-// 	@Prop()
-// 	delay?: any;
+	/**
+	 * Whether or not the popper should size itself to the same width as the
+	 * trigger. Useful for poppers that work like "select" type controls.
+	 */
+	@Prop(propOptional(Boolean))
+	trackTriggerWidth?: boolean;
 
-// 	@Prop(Boolean)
-// 	disabled?: boolean;
+	/**
+	 * Whether or not the popper should take up the full max width instead of
+	 * relying on its content to size itself. Useful for poppers that change the
+	 * content dynamically and you want it to stay one consistent size.
+	 */
+	@Prop(propOptional(Boolean))
+	forceMaxWidth?: boolean;
 
-// 	@Prop(Boolean)
-// 	show?: boolean;
+	/**
+	 * Delay for a hover-based popper showing and hiding. 'show' delay currently defaults
+	 * to TransitionTime if there is no delay specified, and 'hide' delay defaults to at
+	 * least 50ms to make sure there are no issues when moving between popper and trigger.
+	 */
+	@Prop(propOptional(Object))
+	delay?: DelayFormat;
 
-// 	@Prop({ type: Boolean, default: true })
-// 	autoHide!: boolean;
+	// We set a watch on this prop so we know when to display 'manual' triggers
+	@Prop(Boolean)
+	show?: boolean;
 
-// 	@Prop(Boolean)
-// 	block?: boolean;
+	// Sets 'display: block !important' on the trigger element.
+	@Prop(propOptional(Boolean))
+	block?: boolean;
 
-// 	@Prop(String)
-// 	openGroup?: string;
+	@Prop(propOptional(String, null))
+	popoverClass!: null | string;
 
-// 	@Prop(String)
-// 	popoverClass?: string;
+	$refs!: {
+		trigger: any;
+		popper: any;
+	};
 
-// 	$refs!: {
-// 		popover: any;
-// 		scroller: AppScrollScrollerTS;
-// 	};
+	isHiding = false;
+	isVisible = false;
+	width = '';
+	maxWidth = '';
+	popperIndex = PopperIndex++;
 
-// 	isVisible = false;
-// 	width = '';
-// 	maxWidth = '';
-// 	popperIndex = PopperIndex++;
+	ResizeObserver!: ResizeObserver | null;
+	popperInstance!: Instance | null;
 
-// 	private hideTimeout?: NodeJS.Timer;
-// 	private mobileBackdrop: AppBackdrop | null = null;
+	private _triggerElement!: HTMLElement;
+	private hideTimeout?: NodeJS.Timer;
+	private mobileBackdrop: AppBackdrop | null = null;
 
-// 	get maxHeight() {
-// 		return Screen.height - 100 + 'px';
-// 	}
+	get maxHeight() {
+		return Screen.height - 100 + 'px';
+	}
 
-// 	get popperId() {
-// 		return 'popper-' + this.popperIndex;
-// 	}
+	get popperId() {
+		return 'popper-' + this.popperIndex;
+	}
 
-// 	get popoverInnerClass() {
-// 		let classes = ['popper-content'];
-// 		if (this.trackTriggerWidth) {
-// 			classes.push('-track-trigger-width');
-// 		}
-// 		if (this.forceMaxWidth) {
-// 			classes.push('-force-max-width');
-// 		}
-// 		return classes.join(' ');
-// 	}
+	get triggerClass() {
+		let classes = [];
 
-// 	get actualTrigger(): ActualTrigger {
-// 		return this.trigger === 'right-click' ? 'manual' : this.trigger;
-// 	}
+		if (this.block) {
+			classes.push('-block');
+		}
 
-// 	mounted() {
-// 		Popper.registerPopper(this.$router, this);
-// 	}
+		if (this.isVisible) {
+			classes.push('popped');
+		}
 
-// 	destroyed() {
-// 		Popper.deregisterPopper(this);
-// 		this.clearHideTimeout();
-// 		this.removeBackdrop();
-// 	}
+		return classes.join(' ');
+	}
 
-// 	hide() {
-// 		this.$refs.popover.hide();
-// 	}
+	get contentClass() {
+		let classes = [this.popoverClass];
 
-// 	@Emit('show')
-// 	onShow() {
-// 		this.clearHideTimeout();
-// 		this.isVisible = true;
+		if (this.trackTriggerWidth) {
+			classes.push('-track-trigger-width');
+		}
 
-// 		// If we are tracking a particular element's width, then we set this popover to be the same
-// 		// width as the element. We don't track width when it's an XS screen since we do a full
-// 		// width popover in those cases.
-// 		let widthElem: HTMLElement | undefined;
-// 		if (this.trackTriggerWidth && !Screen.isWindowXs) {
-// 			widthElem = document.querySelector('#' + this.popperId) as HTMLElement | undefined;
-// 			if (widthElem) {
-// 				this.width = widthElem.offsetWidth + 'px';
-// 				this.maxWidth = 'none';
-// 			}
-// 		}
+		if (this.forceMaxWidth) {
+			classes.push('-force-max-width');
+		}
 
-// 		// If no element to base our width on, reset.
-// 		if (!widthElem) {
-// 			this.maxWidth = '';
-// 			this.width = '';
-// 		}
+		return classes.join(' ');
+	}
 
-// 		this.addBackdrop();
-// 	}
+	get popperOptions(): Options {
+		return {
+			placement: this.placement,
+			modifiers: [...modifiers],
+			strategy: this.fixed ? 'fixed' : 'absolute',
+		};
+	}
 
-// 	onHide() {
-// 		this.clearHideTimeout();
-// 		this.hideTimeout = setTimeout(() => this.hideDone(), TransitionTime);
-// 		this.removeBackdrop();
-// 	}
+	get actualDelay(): DelayFormat {
+		// Need at least 50ms delay for 'hide' to prevent any
+		// timing issues when moving between trigger and popper.
+		if (this.delay) {
+			return {
+				show: Math.max(this.delay.show, 0),
+				hide: Math.max(this.delay.hide, 50),
+			};
+		}
 
-// 	private addBackdrop() {
-// 		if (Screen.isXs && !this.mobileBackdrop) {
-// 			this.mobileBackdrop = Backdrop.push({ className: 'popper-backdrop' });
-// 		}
-// 	}
+		return { show: TransitionTime, hide: 50 };
+	}
 
-// 	private removeBackdrop() {
-// 		if (Screen.isXs && this.mobileBackdrop) {
-// 			Backdrop.remove(this.mobileBackdrop);
-// 			this.mobileBackdrop = null;
-// 		}
-// 	}
+	mounted() {
+		Popper.registerPopper(this.$router, this);
+		this._triggerElement = this.$el as HTMLDivElement;
 
-// 	@Emit('hide')
-// 	private hideDone() {
-// 		this.isVisible = false;
-// 	}
+		if (this.trigger === 'hover') {
+			this._triggerElement.addEventListener('mouseenter', this.onMouseEnter);
+			this._triggerElement.addEventListener('mouseleave', this.onMouseLeave);
+		}
+	}
 
-// 	private clearHideTimeout() {
-// 		if (this.hideTimeout) {
-// 			clearTimeout(this.hideTimeout);
-// 			this.hideTimeout = undefined;
-// 		}
-// 	}
+	destroyed() {
+		Popper.deregisterPopper(this);
+		this.clearHideTimeout();
+		this.removeBackdrop();
 
-// 	@Emit('auto-hide')
-// 	private onAutoHide() {}
+		// Destroy any poppers that had their trigger element removed,
+		// removing things like floating 'hover' poppers on state change.
+		if (this.popperInstance) {
+			this.popperInstance.destroy();
+		}
+	}
 
-// 	private onContextMenu(e: MouseEvent) {
-// 		if (this.trigger !== 'right-click') {
-// 			return;
-// 		}
+	triggerClicked() {
+		// We want to prevent right-click, hover, and manual triggers from showing poppers on left-click.
+		// clickAway() listener will hide poppers when needed, so we only need to show poppers here.
+		if (this.trigger !== 'click' || this.isVisible) {
+			return;
+		}
 
-// 		e.preventDefault();
-// 		Popper.hideAll();
-// 		this.$refs.popover.show();
-// 	}
-// }
+		this.onShow();
+	}
+
+	onContextMenu(event: MouseEvent) {
+		if (this.trigger !== 'right-click') {
+			return;
+		}
+
+		if (this.isVisible) {
+			return event.preventDefault();
+		}
+
+		event.preventDefault();
+		Popper.hideAll();
+		this.onShow();
+	}
+
+	private clickAway(event: MouseEvent) {
+		if (
+			this.$refs.popper.contains(event.target) ||
+			(this.$refs.trigger.contains(event.target) && this.trigger === 'manual')
+		) {
+			return;
+		}
+
+		this.onHide();
+		document.removeEventListener('click', this.clickAway, true);
+	}
+
+	onMouseEnter() {
+		this.clearHideTimeout();
+
+		// We need to skip the hideTimeout in 'onMouseLeave()' if the
+		// user hovers while the popper is in the process of hiding.
+		if (this.isHiding) {
+			this.hideDone();
+		}
+
+		if (this.isVisible) {
+			return;
+		}
+
+		this.hideTimeout = setTimeout(() => this.onShow(), this.actualDelay.show);
+	}
+
+	onMouseLeave() {
+		this.clearHideTimeout();
+
+		if (!this.isVisible) {
+			return;
+		}
+
+		this.hideTimeout = setTimeout(() => {
+			this.onHide();
+		}, this.actualDelay.hide);
+	}
+
+	async createPopper() {
+		this.isVisible = true;
+		await this.$nextTick();
+
+		this.popperInstance = createPopper(
+			this._triggerElement,
+			this.$refs.popper,
+			this.popperOptions
+		);
+
+		this.ResizeObserver = new ResizeObserver(() => {
+			if (this.popperInstance) {
+				this.popperInstance.update();
+			}
+		});
+		this.ResizeObserver.observe(this.$refs.popper);
+
+		document.body.appendChild(this.$refs.popper);
+
+		if (this.trigger !== 'hover') {
+			document.addEventListener('click', this.clickAway, true);
+		} else {
+			this.$refs.popper.addEventListener('mouseenter', this.onMouseEnter);
+			this.$refs.popper.addEventListener('mouseleave', this.onMouseLeave);
+		}
+	}
+
+	@Emit('show')
+	onShow() {
+		this.clearHideTimeout();
+		this.createPopper();
+
+		// If we are tracking a particular element's width, then we set this popover to be the same
+		// width as the element. We don't track width when it's an XS screen since we do a full
+		// width popover in those cases.
+		let widthElem: HTMLElement | undefined;
+		if (this.trackTriggerWidth && !Screen.isWindowXs) {
+			widthElem = this._triggerElement as HTMLElement | undefined;
+			if (widthElem) {
+				this.width = widthElem.offsetWidth + 'px';
+				this.maxWidth = 'none';
+			}
+		}
+
+		// If no element to base our width on, reset.
+		if (!widthElem) {
+			this.maxWidth = '';
+			this.width = '';
+		}
+
+		this.addBackdrop();
+	}
+
+	onHide() {
+		// In case a popper was hidden from something other than a click,
+		// like right-clicking a cbar item or Popover.hideAll() being triggered.
+		if (this.trigger !== 'hover') {
+			document.removeEventListener('click', this.clickAway, true);
+		}
+
+		this.isHiding = true;
+		this.clearHideTimeout();
+		this.hideTimeout = setTimeout(() => this.hideDone(), TransitionTime);
+		this.removeBackdrop();
+	}
+
+	@Emit('hide')
+	private hideDone() {
+		// Making sure that popper doesn't keep tracking positioning
+		if (this.popperInstance) {
+			this.popperInstance.destroy();
+			this.popperInstance = null;
+		}
+		// or keep watching
+		if (this.ResizeObserver) {
+			this.ResizeObserver.unobserve(this.$refs.popper);
+			this.ResizeObserver = null;
+		}
+
+		this.isVisible = false;
+		this.isHiding = false;
+	}
+
+	private addBackdrop() {
+		if (Screen.isXs && !this.mobileBackdrop) {
+			this.mobileBackdrop = Backdrop.push({ className: 'popper-backdrop' });
+		}
+	}
+
+	private removeBackdrop() {
+		if (Screen.isXs && this.mobileBackdrop) {
+			Backdrop.remove(this.mobileBackdrop);
+			this.mobileBackdrop = null;
+		}
+	}
+
+	private clearHideTimeout() {
+		if (this.hideTimeout) {
+			clearTimeout(this.hideTimeout);
+			this.hideTimeout = undefined;
+		}
+	}
+
+	@Watch('show')
+	onManualShow() {
+		if (this.trigger !== 'manual') {
+			return;
+		}
+
+		if (this.show) {
+			return this.onShow();
+		}
+
+		return this.onHide();
+	}
+}
